Refactor kevent() to make the flow of control more obvious.
Add reference counting to the knote structure. This is slightly buggy and
probably unnecessary.

Eliminate the kqueue_lock() and kqueue_unlock() functions, and create
a heirarchy of two locks: the knotelist lock at the filter level, and
a per-knote lock. This needs to be looked at more closely.

Make note of a race condition when kev.flags contains EV_ADD | EV_DISABLE
and a new knote is created.



git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@411 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7