Expose `with sexp` for the Uri types.
This isn't quite the ideal representation yet due to a bug in
Sexplib when used outside Core.  The `Sexplib.Conv` modules that
contain the type definitions for 'sexp_option' and 'sexp_list' are
not exposed in the `Sexplib.Std` module, which leads to build
failures when using those options.

I've worked around the sexp_option failure for now, and this at
least unblocks the sexp representation of Uri.  A future version
will have a more succinct representation for small Uri strings.