Windows: fix socket check
`fstat` on a socket is poor way to test if a handle is a socket.  Use
`GetFileType` instead.  On debug builds, especially on newer versions, we would
trigger an assert.  This addresses the TODO in the code path with an alternative
and allows the previously failing test to pass as well.