Deal with inotify events containing filenames correctly
From inotify(7)

The name field is present only when an event is returned for a file inside a watched directory; it identifies the file pathname relative to the watched directory. This pathname is null-terminated, and may include further null bytes ('\0') to align subsequent reads to a suitable address boundary.

Without this patch, the abort is triggered after the loop exits.  Removing the abort would cause misaligned reads, where the next time get_one_event was called, the structure would be populate with the name of the file from the previous event.