CPACK_COMPONENT_HEADERS_DEPENDS should list components, not groups
With a group here, CPack thinks there's a component called "main",
as well as a group called "main". It then builds packages for the
"missing" component because it isn't included in any groups. This
"component" package is both broken, and built after the main
package. It has the same name as the proper package, so overwrites
it.