FIX Abb scheduler kqueue error on close
The core issue was that aborts happen in the wrong order: we want to
abort the most-dependent futures first and the last dependent either,
with the pivot being around the future that is being aborted.

This showed itself in code which looks like:

abort (with_finally (fun () -> read_that_waits fd) ~finally:(fun () ->
close fd))

The close would be called first, since finally uses the "abort"
feature to cleanup, and then the read would be aborted after.  But the
read code aborts by putting the fd in the "remove from reads" kqueue
change list, and "close" removes the fd from that list as it knows the
fd doesn't exist anymore.  So the order is important.  This has to go
in the order that the read abort is executed first so that the fd goes
in the remove list and then the close can clean that up.

When the abort code was implemented this was a question: does the
order matter?  At the time it wasn't clear.  Upon this bug, it does
seem like there is a real argument to be made that it does: it should
go in most dependent to least.