Return failed changelist entries (and EV_RECEIPT entries) in the event list closes mheily#108 closes mheily#109
Fix return code to indicate that there are changelist entries in the kevent list.

Use pointers to represent current position in kevent list (reduces possibility of errors).

Calculate nevents earlier so the correct number is passed into kevent_copyin.

Have kevent_copyin set the EV_ERROR flags in the kevent it produces, and set errno to 0 if it added a kevent successfully.