Move processing EV_DISPATCH and EV_ONESHOT to the copyout functions of filters themselves
The issue here is there's no way to do this in common code unless we take a performance hit by allocating a temporary array which associate knotes with kevents.

We could do this at the platform level, but then we still have an issue with filters that produce multiple kevents per copyout call.