Make knote mutexes recursive
This is so that disable and delete can be called from within copyout functions, as the copyout functions should really be locking this mutex as well.

There's also not a whole lot of point in a read/write lock here as it's very unlikely there'll be multiple simultaneous readers on the same knote.  You could even argue it's wrong, as it could theoretically allow the same event to be raised multiple times for filters like posix/proc if 2+ threads called kevent simultaneously on the same kqueue.