FIX Terrat Unlock now unlocks WMs in all states
Prior to this change, an unlock only unlocked completed work manifests
or merged PRs.  But one could get a repo into a state where a
repository had a apply or a plan that was stuck in queued or running
etc.  This makes unlock more powerful in that it unlocks a WM or PR in
ANY state.  This is a bit scary but it also means we can get out of
any bad scenario.

Because WMs in these states are not completed, this means we have to
modify unlock to unlock based on a WMs *created_at*.