REFACTOR Terrat Move plan diff into backend
This reduces the amount of data the action has to send back to us,
which is good because we will be storing that output in the backend.
Currently the action sends us the plan text AND the diff
representation of the plan text.  By moving the diff transformation
into the backend we can send half the data.

This is also prep for handling big comments where we want to store all
of the output text in the database for making a GUI.