FIX Terrat Move API calls outside of db tx
If there are TONS of dirspaces, this tx runs way too long.  It also
had the issue where if the action runs before the tx is done, the work
manifest will see the run as not running (queued).

This does have the issue of the whole thing is not a tx so if
something fails midway through we might be left with a half-setup run.
We should be able to "unlock" our way out of this, though.