PRO-134 FIX Terrat Support repository rename
There was a bug because a repository rename does not have a conflict
on the owner/name but it does on the id.  This switches over so any
conflicting id will update everything.  The assumption is that the id
is the most consistent thing.  There is still a bug in that if a repo
is deleted and created again, it might get a new id, which would not
violate the id restriction but would the owner/name restriction.