PRO-192 FIX Aborts combined with multiple triggered fds
If two fd's are triggered, 1 and 2, at the same time, and each one
aborts the other, we ran into issues where we would either try to
process the other one after it was aborted (and fail) or if we closed
the fd and opened another one, we would call the wrong handler.  So
intead, if we close, or otherwise abort, and fd then we add it to the
ignore list so we know for the rest of the run that we will not
process it.  After each iteration, we clean up the ignore list.