PRO-178 PRO-201 FIX Use merged hash for all operations
GitHub creates a merge for us so we can see if the change is mergable.
We should have been using this hash everywhere but we were only using
it in some places because of how our API works and the GitHub API
works.  That is, we need to use the hash for the HEAD of the users
branch for verification that we are running on the correct change when
the action runs, because we can only start GitHub Actions with branch
names.  But we actually want to run the change on the merge.  And our
generic event evaluator cared about which hash it was using (for no
good reason) so we were stuck using the branch HEAD hash for a bunch
of things that we really wante to use the merge hash for.

So we refactored the generic evaluator to care a lot less about
specific hashes and instead specify semantically what it wants done
(does it want a repo config from the base branch or the "change"?)