PRO-230 FIX Retry fetching pull request if mergeable_state is unknown
Every now and then we see an issue with the GitHub API where we when
we query a pull request, we see the correct HEAD but the merge sha
that GitHub generates is for the previous version of the PR, they have
not generated a new one yet.

The hope is that by retrying we will get the new one, but we're not
sure if it actually works.  There does seem to be a false-positive
issue where we the mergeable_state is unknown so we retry but the
merge sha is correct.