PRO-431 FIX Abort duplicate plans and only run the latest
Note that if a user quickly pushes a new commit, this will work,
however it might look confusing in the logs because it will look like
the work manifest does not exist, when in reality that is because it
is looking for the work manifest based on the SHA and work manifest
id, but the commit has changed do to the push.  This has the same
effect as the aborted work manifest being found, but it is confusing
in the logs.