PRO-452 FIX Do not necessarily autoplan in case of draft PR made ready
Instead, determine if there are any missing plans for that commit and
run those, otherwise do nothing.  This fits users' expectations better
in that if they have been planning (or auto planning) their changes in
draft the whole time, then they do not expect switching to ready will
cause a whole new autoplan.

It is not clear what users expect if they have no planned all of the
changes in this case.  This change only plans the missing dirspaces,
but it could be that if there are any missing ones, users expect
everything to be planned, that way it all shows up in the same comment
in the PR.