PRO-627 FIX When forking a future, make wrapped future a dependency
Forking is used to be able to evaluate multiple futures concurrently.
fork wraps a future, returning a new future that is evaluated once the
underlying future is forked.  This future can be aborted, however it
was a black hole because it had no connection to the underlying future
running.  This adds the underlying future as a dependency to the
returned future when forking.  Now whole graphs of operations can be
aborted.