PRO-627 ADD protect_finally for safe operations
There are some operations that erquire a setup to be run to completion
and then teardown done if the operation was aborted during setup or
after the body is run.  protect_finally enabled a pattern of setup
(that cannot be aborted) -> body -> finally.  If an abort is performed
while setup, it is executed after setup is completed: setup ->
finally.