#212 FIX Source plan from correct hash on merged PR
When is the last merged PR, we correctly use the "merged" hash for
creating and sourcing plans.  However, when the PR is merged but NOT
the most recently merged, we were looking for the plan for the
"merged" ref when in reality we should look at it for HEAD.  That way
if HEAD changes, a re-plan is required.  This issue would have allowed
some cases where a plan was executed that did not correspond to the
code (although it would still be the plan that was approved).