#885 FIX Runs pagination
Remove the unique id check.  This is not necessary in pagination
because the cursor guarantees that the next page is after the existing
one.  Additionally, using the work manifest id does not work here
because this is a "dirspace" centric view, so multiple rows in the
output have the same work manfiest id, so we were actually removing
valid duplication in this check.