#976 FIX Default stack was missing if not explicitly configured
This also addresses an issue of stacks looking quite confused.  We
were trying to reify the stack view just using the repo topology, but
there isn't actually enough information to properly do that, so
instead we save explicit stack information in the synthesized
configuration that can be used to construct the stacks view properly.

There is still some oddities around the ordering.  For example, if
there is a stack with a bunch of dirspaces but they have no rules,
then the part of the system deciding which order to run specific
dirspaces doesn't respect what stack they are part of.  This probably
requires a hefty refactoring to make work.